Fix padding of lpddr4 ram trainigs binaries
The buildroot code for boot image generation padded all lpddr4 ram
trainigs binaries. Because of this the boot image did get to big für the
IRAM.

The NXP-Yocto code does not pad the last binarie so the bootiamge can
fit into the IRAM. We now do the same in Buildroot.