Update bluez_utils package
bluez_utils was backportet to 5.39 to be compatible with the silex
driver.
As fsimx8mm does not use the silex chip we update to the current version
and remove the silex specific packages.

We will have to check for new patches when updating architectures which
are using the silex chip