Add bulding of install.scr script with buildroot
Until now the install.scr script was built externally with a own mkimage
tool.

For better reproducibility, we want to build the install.txt with
Buildroot/Yocto.

We use the mkimage tool from the host-uboot-tools package, which might
have a different version than the uboot we build but should still work.

The install script is executed in finalscript_min-common so it will be
applied for every build. By default the source at
board/f+s/common/install.txt is used.

If one wants to use a different source, they can st the variable
$INSTALL_TXT_PATH in the board specific final script to the respective
source.

Also remove the imx8mm specific version, because it should work with the
common one.