Add package mxm-wifiex-fs for AW CM276NF WLAN chip
F&S uses the Azurewave AW CM276NF WLAN module on several boards. Up to
now we used the mwifiex driver from the Linux kernel. However there is
a newer driver mxm-wifiex available from NXP (ex-Marvell) that will
support more WLAN chips in the future in one driver. It has also
proper support for the 802.11d protocol to switch the regulatory domain
based on the country code that is sent by the access point it is
associated with.

Add package mxm-wifiex-fs that downloads the new NXP driver and builds
the two kernel modules mlan.ko and moal.ko. Optionally the mlanutl
userspace program can be added that can show and set low-level WLAN
chip settings. This may help when certifying the final product.

The package is already prepared to activate more WLAN chips in the
driver, but we do not use any of these on our boards and there is no
proper configuration and calibration data included for them. So only
SD8997 makes sense at the moment. This chip is used on the AW CM276NF.

The package also adds a WLAN configuration file and calibration data
for AW CM276NF. The calibration data is taken from the file

  wlan_src/mapp/mlanconfig/config/cal_data.conf

from the Azurewave driver package

  SD-WLAN-SD-BT-8997-U16-MMC-W16.68.10.p162-16.26.10.p162-C4X16693_V4-MGPL.tar

At runtime, the driver offers some useful data in /proc/mwlan/. By issuing
commands to the config file there it can even be switched to test mode and
specific tests can be run. See the documentation in

  mxm_wifiex/wlan_src/README_MLAN

in the driver source code.

There is currently no Bluetooth support included in this package and
we did not test if the old Bluetooth driver from the kernel can work
together with this new mxm-wifiex WLAN driver. So for now we assume that
the following settings in the Linux kernel are switched off:

  CONFIG_BT_MRVL
  CONFIG_BT_MRVL_SDIO

And of course the old WLAN driver should be switched off, too:

  CONFIG_MWIFIEX
  CONFIG_MWIFIEX_SDIO