Add mxm-wifiex.conf to let udev/modprobe handle module parameters
Now that we are on kmod, add an mxm-wifiex.conf file to the mxm-wifiex-fs
package. This file is copied to etc/modprobe.d. It is used by udev and
modprobe when loading the moal.ko kernel module. We use this to pass the
configuration file argument (mod_para) to the moal.ko module.