Fix GTK3 build on X11
When compiling the xserver with GTK enabled, also the glamor subdirectory
is compiled with the GBM variant enabled. However ompilation of this
directory shows a conflict for struct GC, which is defined in two different
include files.

Investigation showed that the variant that is pulled in via xlib.h is not
wanted in this case. The code actually sets macro MESA_EGL_NO_X11_HEADERS
to achieve this, but as the name already indicates, this only works if
these headers come from the Mesa3D package. However in case of i.MX6, the
X11 headers come from the Vivante graphics libraries. Fortunately, these
headers also provide a way to exclude the appropriate part. But here it
is macro __GBM__ that needs to be set.

The major part of the include files is nonetheless taken from the Mesa3D
package. So add a patch to package/mesa3d that tells pkg-config to also
define this __GBM__ macro. This has the result that the automatically
created Makefiles in the xserver package will have this added to their
GBM_CFLAGS macro.

Unfortunately, GBM_CFLAGS is not used anywhere in the xserver build
process. So also add a patch to package/x11r7/xserver_xorg-server/1.20.5
that adds the GBM_CFLAGS to AM_FLAGS in the glamor subdirectory. This ends
up in adding them to the CFLAGS value that is actually used in compilation.