Add dmix as default for asound.conf
To output multiple audio streams simultaneously, the dmix must be used
to merge them into one output stream. The dmix should always be
available for the alsa packages in Buildroot.
To ensure that the dmix output is listed in higher level audio managers,
the hint -> description needs to be present.