Fix alsactl restore
alsactl restore did not work anymore after the update to br2022.02.
To fix this we have to call the alsactl restore command later in the
init process.

Also we  have to create the /run/lock/ directory because alsactl uses it
to lock its actions.

Also remove ALSA_LIB_UCM because we don't use it and it throws an error
when using alsactl restore

Finally, update asound.state to the yocto version.