Fix qt5 webengine
With the "-webengine-icu" config option qt5webengine uses icu headers
on the build host. If (for ubuntu) libicu-dev is not installed on the
system, build fails with:

In file included from
../../3rdparty/chromium/v8/src/regexp/gen-regexp-special-case.cc:10:
../../3rdparty/chromium/v8/src/regexp/special-case.h:12:10: fatal error:
unicode/uchar.h: No such file or directory
   12 | #include "unicode/uchar.h"
      |          ^~~~~~~~~~~~~~~~~
compilation terminated.

Otherwise, if (for ubuntu) libicu-dev is installed on the system, build
fails later at linking stage.

Using icu bundled with qt solves this problem

https://lists.buildroot.org/pipermail/buildroot/2023-October/677826.html

qt5webengine fails to build due to changes introduced in python3.11:
https://lore.kernel.org/buildroot/20231025205233.1925727-1-kadir.c.yilmaz@gmail.com/T/