Improve u-boot install-script gerneration
We want to be able to build the install/update scripts out of tree. So
we have to adapt the output path if the $O variable is set.

Also rework the mkinstall script so we can build specific update and
install scripts for each architecture if needed.

Update scripts will be architecture specific most of the time so move
them to the architecture specific directory